requirements.txt
# Core AI/ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
torch==2.0.1
torchvision==0.15.2

# Time Series & Forecasting
statsmodels==0.14.0
pmdarima==2.0.4
prophet==1.1.5

# Data Processing
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Network & Simulation Integration
requests==2.31.0
pyyaml==6.0
python-dotenv==1.0.0
httpx==0.24.1

# Database & Logging
sqlalchemy==2.0.20
pymongo==4.4.1
python-dateutil==2.8.2

# Performance & Monitoring
psutil==5.9.5
py-cpuinfo==9.0.0

# Jupyter & Notebooks
jupyter==1.0.0
jupyterlab==4.0.4
ipython==8.14.0
ipywidgets==8.0.7

# Utilities
tqdm==4.66.1
click==8.1.7
colorama==0.4.6
pydantic==2.2.1
joblib==1.3.2          # EnergyPredictor model persistence

# API & Web
fastapi==0.103.0
uvicorn==0.23.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0
